+++ css/style.css
@@ -1416,3 +1416,13 @@ main {
 .hidden {
   display: none;
 }
+
+.input-error {
+  outline: none !important;
+  border: 2px solid tomato;
+}
+
+.input-error:focus {
+  outline: 1px solid tomato !important;
+  border: 1px solid tomato;
+}
