+++ js/api.js
@@ -1,0 +1,32 @@
+const getData = (onSuccess, onFail) => {
+  fetch('https://24.javascript.pages.academy/kekstagram/data')
+    .then((response) => response.json())
+    .then((photos) => {
+      onSuccess(photos);
+    })
+    .catch(() => {
+      onFail('Не удалось получить изображения');
+    });
+};
+
+const sendData = (onSuccess, onFail, body) => {
+  fetch(
+    'https://24.javascript.pages.academy/kekstagram',
+    {
+      method: 'POST',
+      body: body,
+    },
+  )
+    .then((response) => {
+      if (response.ok) {
+        onSuccess();
+      } else {
+        onFail('Не удалось отправить форму. Попробуйте ещё раз');
+      }
+    })
+    .catch(() => {
+      onFail('Не удалось отправить форму. Попробуйте ещё раз');
+    });
+};
+
+export {getData, sendData};
